+++ tests/test_geometries.py
@@ -1,12 +1,8 @@
-import re
-from typing import Union
-
 import pytest
+import shapely
 from pydantic import ValidationError
-from shapely.geometry import shape
 
 from geojson_pydantic.geometries import (
-    Geometry,
     GeometryCollection,
     LineString,
     MultiLineString,
@@ -35,13 +31,6 @@ def test_pydantic_schema(obj):
     assert obj.model_json_schema()
 
 
-def assert_wkt_equivalence(geom: Union[Geometry, GeometryCollection]):
-    """Assert WKT equivalence with Shapely."""
-    # Remove any trailing `.0` to match Shapely format
-    clean_wkt = re.sub(r"\.0(\D)", r"\1", geom.wkt)
-    assert shape(geom).wkt == clean_wkt
-
-
 @pytest.mark.parametrize("coordinates", [(1.01, 2.01), (1.0, 2.0, 3.0), (1.0, 2.0)])
 def test_point_valid_coordinates(coordinates):
     """
@@ -51,7 +40,6 @@ def test_point_valid_coordinates(coordinates):
     assert p.type == "Point"
     assert p.coordinates == coordinates
     assert hasattr(p, "__geo_interface__")
-    assert_wkt_equivalence(p)
 
 
 @pytest.mark.parametrize(
@@ -88,7 +76,6 @@ def test_multi_point_valid_coordinates(coordinates):
     assert p.type == "MultiPoint"
     assert p.coordinates == coordinates
     assert hasattr(p, "__geo_interface__")
-    assert_wkt_equivalence(p)
 
 
 @pytest.mark.parametrize(
@@ -123,7 +110,6 @@ def test_line_string_valid_coordinates(coordinates):
     assert linestring.type == "LineString"
     assert linestring.coordinates == coordinates
     assert hasattr(linestring, "__geo_interface__")
-    assert_wkt_equivalence(linestring)
 
 
 @pytest.mark.parametrize("coordinates", [None, "Foo", [], [(1.0, 2.0)], ["Foo", "Bar"]])
@@ -162,7 +148,6 @@ def test_multi_line_string_valid_coordinates(coordinates):
     assert multilinestring.type == "MultiLineString"
     assert multilinestring.coordinates == coordinates
     assert hasattr(multilinestring, "__geo_interface__")
-    assert_wkt_equivalence(multilinestring)
 
 
 @pytest.mark.parametrize(
@@ -200,7 +185,6 @@ def test_polygon_valid_coordinates(coordinates):
     else:
         assert polygon.exterior is None
     assert not list(polygon.interiors)
-    assert_wkt_equivalence(polygon)
 
 
 @pytest.mark.parametrize(
@@ -248,7 +232,6 @@ def test_polygon_with_holes(coordinates):
     assert hasattr(polygon, "__geo_interface__")
     assert polygon.exterior == polygon.coordinates[0]
     assert list(polygon.interiors) == [polygon.coordinates[1]]
-    assert_wkt_equivalence(polygon)
 
 
 @pytest.mark.parametrize(
@@ -325,7 +308,6 @@ def test_multi_polygon(coordinates):
 
     assert multi_polygon.type == "MultiPolygon"
     assert hasattr(multi_polygon, "__geo_interface__")
-    assert_wkt_equivalence(multi_polygon)
 
 
 @pytest.mark.parametrize(
@@ -494,7 +476,6 @@ def test_geometry_collection_iteration(coordinates):
         type="GeometryCollection", geometries=[polygon, multipolygon]
     )
     assert hasattr(gc, "__geo_interface__")
-    assert_wkt_equivalence(gc)
     iter(gc)
 
 
@@ -508,7 +489,6 @@ def test_len_geometry_collection(coordinates):
     gc = GeometryCollection(
         type="GeometryCollection", geometries=[polygon, multipolygon]
     )
-    assert_wkt_equivalence(gc)
     assert len(gc) == 2
 
 
@@ -522,7 +502,6 @@ def test_getitem_geometry_collection(coordinates):
     gc = GeometryCollection(
         type="GeometryCollection", geometries=[polygon, multipolygon]
     )
-    assert_wkt_equivalence(gc)
     assert polygon == gc[0]
     assert multipolygon == gc[1]
 
@@ -530,13 +509,19 @@ def test_getitem_geometry_collection(coordinates):
 def test_wkt_mixed_geometry_collection():
     point = Point(type="Point", coordinates=(0.0, 0.0, 0.0))
     line_string = LineString(type="LineString", coordinates=[(0.0, 0.0), (1.0, 1.0)])
-    gc = GeometryCollection(type="GeometryCollection", geometries=[point, line_string])
-    assert_wkt_equivalence(gc)
+    assert (
+        GeometryCollection(
+            type="GeometryCollection", geometries=[point, line_string]
+        ).wkt
+        == "GEOMETRYCOLLECTION Z (POINT Z (0.0 0.0 0.0), LINESTRING (0.0 0.0, 1.0 1.0))"
+    )
 
 
 def test_wkt_empty_geometry_collection():
-    gc = GeometryCollection(type="GeometryCollection", geometries=[])
-    assert_wkt_equivalence(gc)
+    assert (
+        GeometryCollection(type="GeometryCollection", geometries=[]).wkt
+        == "GEOMETRYCOLLECTION EMPTY"
+    )
 
 
 def test_geometry_collection_warnings():
@@ -734,3 +719,45 @@ def test_wkt_empty_geometrycollection():
     assert GeometryCollection(type="GeometryCollection", geometries=[]).wkt.endswith(
         " EMPTY"
     )
+
+
+@pytest.mark.parametrize(
+    "wkt",
+    (
+        "POINT (0.0 0.0)",
+        # "POINT EMPTY" does not result in valid GeoJSON
+        "POINT Z (0.0 0.0 0.0)",
+        "MULTIPOINT ((0.0 0.0))",
+        "MULTIPOINT Z ((0.0 0.0 0.0))",
+        "MULTIPOINT ((0.0 0.0), (1.0 1.0))",
+        "MULTIPOINT Z ((0.0 0.0 0.0), (1.0 1.0 1.0))",
+        "MULTIPOINT EMPTY",
+        "LINESTRING (0.0 0.0, 1.0 1.0, 2.0 2.0)",
+        "LINESTRING Z (0.0 0.0 0.0, 1.0 1.0 1.0, 2.0 2.0 2.0)",
+        # "LINESTRING EMPTY" does not result in valid GeoJSON
+        "MULTILINESTRING ((0.0 0.0, 1.0 1.0))",
+        "MULTILINESTRING ((0.0 0.0, 1.0 1.0), (1.0 1.0, 2.0 2.0))",
+        "MULTILINESTRING Z ((0.0 0.0 0.0, 1.0 1.0 1.0))",
+        "MULTILINESTRING Z ((0.0 0.0 0.0, 1.0 1.0 1.0), (1.0 1.0 1.0, 2.0 2.0 2.0))",
+        "MULTILINESTRING EMPTY",
+        "POLYGON ((0.0 0.0, 1.0 1.0, 2.0 2.0, 3.0 3.0, 0.0 0.0))",
+        "POLYGON ((0.0 0.0, 4.0 0.0, 4.0 4.0, 0.0 4.0, 0.0 0.0), (1.0 1.0, 1.0 2.0, 2.0 2.0, 2.0 1.0, 1.0 1.0))",
+        "POLYGON Z ((0.0 0.0 0.0, 1.0 1.0 0.0, 2.0 2.0 0.0, 3.0 3.0 0.0, 0.0 0.0 0.0))",
+        "POLYGON Z ((0.0 0.0 0.0, 4.0 0.0 0.0, 4.0 4.0 0.0, 0.0 4.0 0.0, 0.0 0.0 0.0), (1.0 1.0 0.0, 1.0 2.0 0.0, 2.0 2.0 0.0, 2.0 1.0 0.0, 1.0 1.0 0.0))",
+        "POLYGON EMPTY",
+        "MULTIPOLYGON (((0.0 0.0, 1.0 1.0, 2.0 2.0, 3.0 3.0, 0.0 0.0)))",
+        "MULTIPOLYGON (((0.0 0.0, 1.0 1.0, 2.0 2.0, 3.0 3.0, 0.0 0.0)), ((1.0 1.0, 2.0 2.0, 3.0 3.0, 4.0 4.0, 1.0 1.0)))",
+        "MULTIPOLYGON Z (((0.0 0.0 0.0, 1.0 1.0 0.0, 2.0 2.0 0.0, 3.0 3.0 0.0, 0.0 0.0 0.0)))",
+        "MULTIPOLYGON Z (((0.0 0.0 0.0, 1.0 1.0 0.0, 2.0 2.0 0.0, 3.0 3.0 0.0, 0.0 0.0 0.0)), ((1.0 1.0 0.0, 2.0 2.0 0.0, 3.0 3.0 0.0, 4.0 4.0 0.0, 1.0 1.0 0.0)))",
+        "MULTIPOLYGON EMPTY",
+        "GEOMETRYCOLLECTION (POINT (0.0 0.0))",
+        "GEOMETRYCOLLECTION (POINT (0.0 0.0), MULTIPOINT ((0.0 0.0), (1.0 1.0)))",
+        "GEOMETRYCOLLECTION Z (POLYGON EMPTY, MULTIPOLYGON Z (((0.0 0.0 0.0, 1.0 1.0 0.0, 2.0 2.0 0.0, 3.0 3.0 0.0, 0.0 0.0 0.0))))",
+        "GEOMETRYCOLLECTION Z (LINESTRING Z (0.0 0.0 0.0, 1.0 1.0 1.0, 2.0 2.0 2.0), MULTILINESTRING ((0.0 0.0, 1.0 1.0), (1.0 1.0, 2.0 2.0)))",
+        "GEOMETRYCOLLECTION EMPTY",
+    ),
+)
+def test_wkt(wkt: str):
+    # Use Shapely to parse the input WKT so we know it is parsable by other tools.
+    # Then load it into a Geometry and ensure the output WKT is the same as the input.
+    assert parse_geometry_obj(shapely.from_wkt(wkt).__geo_interface__).wkt == wkt